perf(parser): use sax-js for better performance
I ran into performance problems where saxes just blocks the whole thing
for a second or two depending on how large is the input.

I noticed that other libraries that use sax-js don't suffer the same
problem so I  decided to give sax-js a shot and boom! the problem is
gone.